requirements.txt
alembic==1.4.3
asyncpg==0.22.0
bcrypt==3.2.0
cachetools==4.2.1
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
databases==0.4.3
Deprecated==1.2.12
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
Faker==4.14.2
fastapi==0.61.2
fastapi-camelcase==1.0.2
fastapi-jwt-auth==0.5.0
fastapi-users==5.1.3
fastapi-utils==0.2.1
future==0.18.2
h11==0.11.0
httpcore==0.12.3
httplib2==0.19.1
httpx==0.17.1
httpx-oauth==0.3.5
idna==2.10
makefun==1.11.3
Mako==1.1.3
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
passlib==1.7.4
protobuf==3.15.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.2
pyhumps==1.6.1
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-keycloak==0.23.0
python-multipart==0.0.5
python-whois==0.7.3
pytz==2021.1
requests==2.25.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.7.2
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.20
sqlalchemy-pagination==0.0.2
starlette==0.13.6
text-unidecode==1.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.12.2
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
